+++ app/src/main/java/woowacourse/payments/ui/newcard/NewCardScreen.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.card.add
+package woowacourse.payments.ui.newcard
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -6,20 +6,18 @@ import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import woowacourse.payments.card.add.components.CardExpirationDateTextField
-import woowacourse.payments.card.add.components.CardHolderNameTextField
-import woowacourse.payments.card.add.components.CardNumberTextField
-import woowacourse.payments.card.add.components.CardPasswordTextField
-import woowacourse.payments.card.add.components.NewCardTopBar
-import woowacourse.payments.card.add.components.PaymentCard
+import woowacourse.payments.ui.newcard.components.CardExpirationDateTextField
+import woowacourse.payments.ui.newcard.components.CardHolderNameTextField
+import woowacourse.payments.ui.newcard.components.CardNumberTextField
+import woowacourse.payments.ui.newcard.components.CardPasswordTextField
+import woowacourse.payments.ui.newcard.components.NewCardTopBar
+import woowacourse.payments.ui.newcard.components.PaymentCard
 
 @Composable
 fun NewCardScreen(
